train: accept quantized encodings
--encoding only took f16|f32, so training on a q4 base meant passing --dit and
--same as explicit paths. Reuse the inference encoding table instead, split into
encoding.h so the config can validate without pulling in ggml.
