+++ src/sa3_pipeline.h
@@ -24,6 +24,7 @@
 // this. (Open question for the docs: revisit a sa3_pipeline.cpp if incremental build time bites.)
 #pragma once
 
+#include "encoding.h"
 #include "gguf_model.h"
 #include "tokenizer.h"
 #include "t5gemma.h"
@@ -295,29 +296,6 @@ inline bool dit_base_is_quantized(const GgufModel& dit) {
             if (ggml_is_quantized(kv.second->type)) return true;
     }
     return false;
-}
-
-// Canonical file-suffix label for an --encoding value, or "" if unrecognised.
-// Suffixes follow the Encoding field in docs/DISTRIBUTION.md (gguf-spec naming, same
-// spelling llama.cpp uses): Q4_K_M / Q5_K_M, not Q4_KM. The compact spellings are
-// accepted as input aliases so older locally-quantized files and commands still work,
-// but only the canonical form is ever written or resolved.
-inline std::string encoding_suffix(const std::string& encoding) {
-    if (encoding == "f32"     || encoding == "F32")     return "F32";
-    if (encoding == "f16"     || encoding == "F16")     return "F16";
-    if (encoding == "q4_k_m"  || encoding == "Q4_K_M")  return "Q4_K_M";
-    if (encoding == "q5_k_m"  || encoding == "Q5_K_M")  return "Q5_K_M";
-    if (encoding == "q5_k"    || encoding == "Q5_K")    return "Q5_K";
-    if (encoding == "q8_0"    || encoding == "Q8_0")    return "Q8_0";
-    if (encoding == "q4_km"   || encoding == "Q4_KM")   return "Q4_K_M";   // alias
-    if (encoding == "q5_km"   || encoding == "Q5_KM")   return "Q5_K_M";   // alias
-    return "";
-}
-
-// Every canonical encoding label, for "what else is available" diagnostics and CLI help.
-inline const std::vector<std::string>& encoding_labels() {
-    static const std::vector<std::string> v = {"F16", "F32", "Q4_K_M", "Q5_K_M", "Q5_K", "Q8_0"};
-    return v;
 }
 
 inline bool ModelPaths::resolve(const std::string& md, const std::string& variant,
